requirements.txt
Django==1.6.5
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-annoying==0.7.9
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
django-phonenumber-field==0.5
django-toolbelt==0.0.1
gnureadline==6.3.3
gunicorn==18.0
ipython==2.1.0
jsonfield==0.9.20
newrelic==2.20.0.17
phonenumbers==6.1.0
plivo==0.9.6
psycopg2==2.5.3
pycrypto==2.6.1
pystache==0.5.4
python-postmark==0.4.1
requests==2.3.0
sqlparse==0.1.11
static==1.0.2
wsgiref==0.1.2
